% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreste-package.R
\docType{package}
\name{agreste-package}
\alias{agreste}
\alias{agreste-package}
\title{agreste: Aide À La Publication Sur Agreste}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Aide à la création des documents qui, après PAO, pourront être publiés sur Agreste. Facilite la cration de tableaux, graphiques, cartes. Automatise la création des chiffres et données.
}
\author{
\strong{Maintainer}: Anaël Delorme \email{anael.delorme@agriculture.gouv.fr} (\href{https://orcid.org/0000-0002-1122-9076}{ORCID})

Authors:
\itemize{
  \item Killian Steunou \email{killian.steunou@agriculture.gouv.fr}
}

Other contributors:
\itemize{
  \item MinAgri-SSP [copyright holder]
}

}
\keyword{internal}
