% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_champ.R
\name{ajouter_champ}
\alias{ajouter_champ}
\title{Ajouter le champ sous un tableau dans une feuille d'un fichier xlsx}
\usage{
ajouter_champ(
  classeur,
  nom_feuille,
  champ,
  col_debut = 2,
  style = agreste::creer_liste_style_excel()$champ,
  fusion = TRUE
)
}
\arguments{
\item{classeur}{workbook créé avec openXLSX}

\item{nom_feuille}{nom de la feuille}

\item{champ}{le champ à ajouter}

\item{col_debut}{la colonne de départ (2eme par défaut)}

\item{style}{le style à appliquer (style "champ" par défaut)}

\item{fusion}{LGL si les cellules sont à fusionner ou pas (TRUE par défaut)}
}
\value{
Rien n'est renvoyé mais la feuille du classeur est modifiée
}
\description{
Cette fonction permet d'ajouter un champ sous un tableau dans une feuille Excel
}
\examples{

# Création d'un classeur avec une feuille contenant iris et une feuille contenant airquality
# Ajout du champ
# Sauvegarde du classeur
# Ouverture du classeur
library(openxlsx)
library(agreste)
mon_classeur <- createWorkbook()

mes_styles <- creer_liste_style_excel(format = "chiffres_et_donnees")

start_ligne = 3
start_col = 2
nb_col = ncol(iris)

ajouter_tableau_excel(classeur = mon_classeur, tableau = iris, nom_feuille = "iris")
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$ligne_titre, rows = start_ligne, cols = start_col:(nb_col+start_col-1))
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$decimal, rows = (start_ligne+1):(nrow(iris)+start_ligne), cols = 2)
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$decimal, rows = (start_ligne+1):(nrow(iris)+start_ligne), cols = 3)
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$decimal, rows = (start_ligne+1):(nrow(iris)+start_ligne), cols = 4)
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$decimal_italique, rows = (start_ligne+1):(nrow(iris)+start_ligne), cols = 5)
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$texte, rows = (start_ligne+1):(nrow(iris)+start_ligne), cols = 6)

ajouter_titre_tableau(classeur = mon_classeur, nom_feuille = "iris", titre = "Fleurs", col_debut = start_col)
ajouter_source(classeur = mon_classeur, nom_feuille = "iris", source = "Ceci est une source quelconque", col_debut = start_col, avec_note = FALSE)
ajouter_champ(classeur = mon_classeur, nom_feuille = "iris", champ = "France métropolitaine", col_debut = start_col)

start_ligne = 3
start_col = 2
nb_col = ncol(airquality)

ajouter_tableau_excel(classeur = mon_classeur, tableau = airquality, nom_feuille = "airquality")
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$ligne_titre, rows = start_ligne, cols = start_col:(nb_col+start_col-1))
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$numerique, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 2)
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$numerique, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 3)
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$decimal, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 4)
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$numerique, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 5)
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$numerique, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 6)
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$numerique, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 7)

ajouter_titre_tableau(classeur = mon_classeur, nom_feuille = "airquality", titre = "Qualité de l'air", col_debut = start_col)
ajouter_note_lecture(classeur = mon_classeur, nom_feuille = "airquality", note = "blablabla", col_debut = start_col)
ajouter_source(classeur = mon_classeur, nom_feuille = "airquality", source = "Eurostat", col_debut = start_col, avec_note = TRUE)
ajouter_champ(classeur = mon_classeur, nom_feuille = "airquality", champ = "Daily air quality measurements in New York, May to September 1973.", col_debut = start_col)


saveWorkbook(wb = mon_classeur, file = "tableau.xlsx", overwrite = TRUE)

browseURL("tableau.xlsx")

}
