% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_tableau_excel.R
\name{ajouter_tableau_excel}
\alias{ajouter_tableau_excel}
\title{Ajouter un tableau dans une feuille excel}
\usage{
ajouter_tableau_excel(
  classeur,
  tableau,
  nom_feuille,
  ligne_debut = 3,
  col_debut = 2
)
}
\arguments{
\item{classeur}{workbook créé avec openXLSX}

\item{tableau}{dataframe ou tibble à insérer dans une nouvelle feuille}

\item{nom_feuille}{nom de la nouvelle feuille}

\item{ligne_debut}{la ligne à laquelle démarre le tableau}

\item{col_debut}{la colonne à laquelle démarre le tableau}
}
\value{
Rien n'est renvoyé mais l'objet classeur a une nouvelle feuille
}
\description{
Cette fonction crée une nouvelle feuille dans le classeur passé en paramètre et y insère un tableau mis en forme.
}
\examples{

# Création d'un classeur avec une feuille contenant iris et une feuille contenant airquality
# Sauvegarde du classeur
# Ouverture du classeur
library(openxlsx)
library(agreste)
mon_classeur <- createWorkbook()

mes_styles <- creer_liste_style_excel()

start_ligne = 3
start_col = 2
nb_col = ncol(iris)

ajouter_tableau_excel(classeur = mon_classeur, tableau = iris, nom_feuille = "iris")
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$ligne_titre, rows = start_ligne, cols = start_col:(nb_col+start_col-1))
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$decimal, rows = (start_ligne+1):(nrow(iris)+start_ligne), cols = 2)
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$decimal, rows = (start_ligne+1):(nrow(iris)+start_ligne), cols = 3)
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$decimal, rows = (start_ligne+1):(nrow(iris)+start_ligne), cols = 4)
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$decimal_italique, rows = (start_ligne+1):(nrow(iris)+start_ligne), cols = 5)
addStyle(wb = mon_classeur, sheet = "iris", style = mes_styles$texte, rows = (start_ligne+1):(nrow(iris)+start_ligne), cols = 6)

start_ligne = 3
start_col = 2
nb_col = ncol(airquality)

ajouter_tableau_excel(classeur = mon_classeur, tableau = airquality, nom_feuille = "airquality")
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$ligne_titre, rows = start_ligne, cols = start_col:(nb_col+start_col-1))
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$numerique, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 2)
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$numerique, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 3)
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$decimal, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 4)
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$numerique, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 5)
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$numerique, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 6)
addStyle(wb = mon_classeur, sheet = "airquality", style = mes_styles$numerique, rows = (start_ligne+1):(nrow(airquality)+start_ligne), cols = 7)

saveWorkbook(wb = mon_classeur, file = "tableau.xlsx", overwrite = TRUE)

browseURL("tableau.xlsx")

}
