% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_titre_primeur.R,
%   R/creer_liste_style_excel.R
\name{ajouter_titre_primeur}
\alias{ajouter_titre_primeur}
\title{Ajouter le titre du primeur dans la première cellule de chaque feuille}
\usage{
ajouter_titre_primeur(classeur, titre, col_debut = 2, fusion = TRUE)

ajouter_titre_primeur(classeur, titre, col_debut = 2, fusion = TRUE)
}
\arguments{
\item{classeur}{workbook créé avec openXLSX}

\item{titre}{le titre à ajouter}

\item{col_debut}{la colonne de départ (2eme par défaut)}

\item{fusion}{LGL si les cellules sont à fusionner ou pas (TRUE par défaut)}
}
\value{
Rien n'est renvoyé mais la feuille du classeur est modifiée

Rien n'est renvoyé mais la feuille du classeur est modifiée
}
\description{
Cette fonction permet d'ajouter le titre en haut de chaque feuille
d'un classeur de type primeur

Cette fonction permet d'ajouter le titre en haut de chaque feuille
d'un classeur de type primeur
}
\examples{
library(openxlsx)
library(agreste)
mon_classeur <- createWorkbook()

mes_styles <- creer_liste_style_excel(format = "primeur")

start_ligne = 5
start_col = 2
nb_col = ncol(iris)

ajouter_tableau_excel(classeur = mon_classeur,
                      tableau = iris,
                      nom_feuille = "iris",
                      ligne_debut = 5)

## Formatage

addStyle(wb = mon_classeur,
         sheet = "iris",
         style = mes_styles$ligne_titre,
         rows = start_ligne,
         cols = start_col:(nb_col+start_col-1))
addStyle(wb = mon_classeur,
         sheet = "iris",
         style = mes_styles$decimal,
         rows = (start_ligne+1):(nrow(iris)+start_ligne),
         cols = 2)
addStyle(wb = mon_classeur,
         sheet = "iris",
         style = mes_styles$decimal,
         rows = (start_ligne+1):(nrow(iris)+start_ligne),
         cols = 3)
addStyle(wb = mon_classeur,
         sheet = "iris",
         style = mes_styles$decimal,
         rows = (start_ligne+1):(nrow(iris)+start_ligne),
         cols = 4)
addStyle(wb = mon_classeur,
         sheet = "iris",
         style = mes_styles$decimal_italique,
         rows = (start_ligne+1):(nrow(iris)+start_ligne),
         cols = 5)
addStyle(wb = mon_classeur,
         sheet = "iris",
         style = mes_styles$texte,
         rows = (start_ligne+1):(nrow(iris)+start_ligne),
         cols = 6)

## Ajout du titre 

ajouter_titre_primeur(classeur = mon_classeur,
                      titre = "Iris")

## Sauvegarde
saveWorkbook(wb = mon_classeur,
             file = "tableau.xlsx",
             overwrite = TRUE)

browseURL("tableau.xlsx")
}
