% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_formatage.R
\name{app_formatage}
\alias{app_formatage}
\title{Ouvrir l'application Shiny d'édition du tableur}
\usage{
app_formatage()
}
\value{
Rien n'est renvoyé, l'appli s'ouvre.
}
\description{
Cette fonction permet d'ouvrir l'interface graphique s'appuyant
sur les fonctions de formatage du package pour éditer des tableaux plus rapidement
}
\examples{

library(agreste)
\dontrun{
app_formatage()
}
}
