% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_rond_proportionnel.R
\name{carte_rond_proportionnel}
\alias{carte_rond_proportionnel}
\title{Créer une carte représentant les volumes en ronds proportionnels}
\usage{
carte_rond_proportionnel(
  df,
  colonne_code_zone_geo,
  colonne_volume,
  titre_carte = "Carte 1",
  sous_titre_carte = "sous titre de la carte",
  titre_legende_volume = "titre de la legende",
  note_sous_carte = "Lecture : bla bla bla <br> <br> Source : Agreste",
  legend.position = "right",
  zone_geographique = "france",
  niveau_representation = "region_metro",
  couleur_fond_carte = "white",
  couleur_limite_zone = "black",
  taille_limite_zone = 0.3,
  couleur_fond_rond = "#E66A64",
  couleur_contour_rond = "darkgray",
  taille_max_rond = 9,
  hauteur_cm = 8,
  largeur_cm = 8,
  save_carte = FALSE,
  path_carte = "carte.png"
)
}
\arguments{
\item{df}{dataframe avec les données à afficher sur la carte}

\item{colonne_code_zone_geo}{nom de la colonne contenant les id de la zone géo : code région, n° de département...}

\item{colonne_volume}{nom de la colonne contenant les données de volume}

\item{titre_carte}{CHAR Titre à afficher en haut de la carte}

\item{sous_titre_carte}{CHAR Sous-titre à afficher en haut de la carte, en dessous du titre}

\item{titre_legende_volume}{CHAR Titre à afficher en haut de la légende}

\item{note_sous_carte}{CHAR Note à afficher sous la carte}

\item{legend.position}{CHAR choit de la position de la carte "right", "left", "bottom", "top"}

\item{zone_geographique}{CHAR indiquant la zone retenue pour la carte : "france", "ue" ou "monde"}

\item{niveau_representation}{CHAR indiquant le niveau de représentation : "pays", "region_metro" ou "departement_metro"}

\item{couleur_fond_carte}{CHAR couleur de fond la carte "nom_couleur" ou "#code_couleur_html"}

\item{couleur_limite_zone}{CHAR couleur des lignes de séparation de zone géographique "nom_couleur" ou "#code_couleur_html"}

\item{taille_limite_zone}{NUM épaisseur des lignes de séparation de zone géographique}

\item{couleur_fond_rond}{CHAR couleur de remplissage des ronds affichés "nom_couleur" ou "#code_couleur_html"}

\item{couleur_contour_rond}{CHAR couleur de contour des ronds "nom_couleur" ou "#code_couleur_html"}

\item{taille_max_rond}{NUM taille du rond correspondant au plus grand volume}

\item{hauteur_cm}{NUM hauteur de la carte en cm}

\item{largeur_cm}{NUM largeur de la carte en cm}

\item{save_carte}{LGL TRUE si export de la carte au format png, FALSE par défaut}

\item{path_carte}{CHAR chemin vers l'image à sauvegarder}
}
\value{
une carte de type ggplot 2
}
\description{
Créer une carte représentant les volumes en ronds proportionnels
}
\examples{
library(agreste)
library(dplyr)
### récupération des données de permis de construire de SITADEL 
permis_construire_region_2020 <- agreste::permis_construire_region \%>\% 
  filter(annee_autorisation == 2020)

### création de la carte et sauvegarde du fichier carte
carte_pc_2020 <- carte_rond_proportionnel(df = permis_construire_region_2020, colonne_code_zone_geo = REG, colonne_volume = PC, titre_carte = "Carte 1", sous_titre_carte = "Nombre de permis de construire délivré en 2020 par région", titre_legende_volume = "Nombre de permis de construire", save_carte = TRUE, path_carte = "carteReg.png")

### ouverture de la carte
browseURL("carteReg.png")

# Pour faire par département :
permis_construire_dep_2020 <- agreste::permis_construire_dep \%>\%
  filter(annee_autorisation == 2020)

carte_pc_2020 <- carte_rond_proportionnel(df = permis_construire_dep_2020, colonne_code_zone_geo = DEP, colonne_volume = PC, niveau_representation = "departement_metro", titre_carte = "Carte 2", sous_titre_carte = "Nombre de permis de construire délivré en 2020 par département", titre_legende_volume = "Nombre de permis de construire", save_carte = TRUE, path_carte = "carteDEP.png", taille_max_rond = 3)

browseURL("carteDEP.png")

}
