% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_excel_depuis_plan.R
\name{creer_excel_depuis_plan}
\alias{creer_excel_depuis_plan}
\title{Crée un classeur et le formate}
\usage{
creer_excel_depuis_plan(
  fichier,
  format,
  col_debut,
  save = TRUE,
  path = "tableau1.xlsx"
)
}
\arguments{
\item{fichier}{fichier .xlsx contenant les détails pour chaque tableau}

\item{format}{CHAR le type de publication : "chiffres_et_donnees" ou "primeur"}

\item{col_debut}{NUM la colonne à laquelle les tableaux démarrent}

\item{save}{LGL TRUE si export du classeur au format xlsx}

\item{path}{CHAR chemin vers le tableau à sauvegarder}
}
\value{
un objet de type workbook formaté
}
\description{
Cette fonction permet de formater tous les tableaux d'un classeur selon le type de publication choisi, simplement à partir d'un plan
}
\examples{
## Exemple

}
