% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_liste_style_excel.R
\name{creer_liste_style_excel}
\alias{creer_liste_style_excel}
\title{Parametrer le tableur excel}
\usage{
creer_liste_style_excel(format = "chiffres_et_donnees")
}
\arguments{
\item{format}{le format choisi : "chiffres_et_donnees" ou "primeur"}
}
\value{
une liste avec tous les styles disponibles (objets openxls)
}
\description{
Cette fonction permet de créer une liste avec tous les styles utiles aux publications Agreste xlsx
}
\examples{
library(agreste)
mes_styles <- creer_liste_style_excel(format = "chiffres_et_donnees")

#exemples
mes_styles$ligne_titre
mes_styles$numerique

}
