% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees_fond_carte.R
\name{donnees_fond_carte}
\alias{donnees_fond_carte}
\title{Créer un tableau avec les données géographiques de fond de carte}
\usage{
donnees_fond_carte(
  zone_geographique = "france",
  niveau_representation = "region_metro"
)
}
\arguments{
\item{zone_geographique}{CHAR indiquant la zone retenue pour la carte : "france", "ue" ou "monde"}

\item{niveau_representation}{CHAR indiquant le niveau de représentation : "pays", "region_metro" ou "departement_metro"}
}
\value{
Dataframe avec les données géographiques des zones de la carte
}
\description{
Créer un tableau avec les données géographiques de fond de carte
}
\examples{
data_geo_france_region_metro <- donnees_fond_carte(zone_geographique = "france", niveau_representation = "region_metro")
data_geo_france_dep_metro <- donnees_fond_carte(zone_geographique = "france", niveau_representation = "departement_metro")
}
