% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taille_colonnes.R
\name{taille_colonnes}
\alias{taille_colonnes}
\title{Applique des tailles de colonne aux tableaux d'un workbook}
\usage{
taille_colonnes(
  classeur,
  liste_type_donnees,
  largeurs,
  format,
  col_debut = 1,
  largeur_max = TRUE
)
}
\arguments{
\item{classeur}{Le classeur excel, objet de type workbook}

\item{liste_type_donnees}{LIST liste de vecteurs contenant
les types de données de chaque colonne}

\item{largeurs}{LIST liste de vecteurs contenant les largeurs colonne,
"auto" pour appliquer des largeurs automatique}

\item{format}{CHAR le format de publication : "primeur" ou "chiffres_et_donnees"}

\item{col_debut}{La colonne à laquelle commencent les tableaux, par défaut 1}

\item{largeur_max}{LGL si le classeur doit respecter la largeur maximale des
règles de diffusion (83 pour un chiffres et données)}
}
\value{
Rien n'est renvoyé mais le classeur est modifié
}
\description{
Cette fonction permet d'appliquer une taille de colonne choisie par
l'utilisateur ou bien automatiquement,
le total étant égal à 83, et les colonnes de texte étant plus larges
que les autres en mode automatique
}
\examples{
library(openxlsx)
library(agreste)

## Création d'un classeur avec deux feuilles et deux tableaux
mon_classeur <- createWorkbook()

ajouter_tableau_excel(classeur = mon_classeur,
                      tableau = iris,
                      nom_feuille = "iris",
                      ligne_debut = 5,
                      col_debut = 2)

ajouter_titre_tableau(classeur = mon_classeur,
                      nom_feuille = "iris",
                      titre = "Données fleurs iris",
                      col_debut = 2,
                      format = "primeur")
ajouter_source(classeur = mon_classeur,
               nom_feuille = "iris",
               source = "Ceci est une source quelconque",
               format = "primeur",
               col_debut = 2,
               avec_note = FALSE,
               avec_titre = FALSE)
ajouter_champ(classeur = mon_classeur,
              nom_feuille = "iris",
              champ = "France métropolitaine",
              format = "primeur",
              col_debut = 2,
              avec_titre = FALSE)

ajouter_tableau_excel(classeur = mon_classeur,
                      tableau = airquality,
                      nom_feuille = "airquality",
                      ligne_debut = 5,
                      col_debut = 2)

ajouter_titre_tableau(classeur = mon_classeur,
                      nom_feuille = "airquality",
                      titre = "Données qualité de l'air",
                      col_debut = 2,
                      format = "primeur")
ajouter_note_lecture(classeur = mon_classeur,
                     nom_feuille = "airquality",
                     note = "blablabla",
                     format = "primeur",
                     col_debut = 2,
                     avec_titre = FALSE)
ajouter_source(classeur = mon_classeur,
               nom_feuille = "airquality",
               source = "Eurostat",
               format = "primeur",
               col_debut = 2,
               avec_note = TRUE,
               avec_titre = FALSE)
ajouter_champ(classeur = mon_classeur,
              nom_feuille = "airquality",
              champ = "Daily air quality measurements in New York, May to September 1973.",
              format = "primeur",
              col_debut = 2,
              avec_titre = FALSE)

ajouter_titre_primeur(classeur = mon_classeur,
                      titre = "Exemple 2022")

taille_colonnes(classeur = mon_classeur,
                liste_type_donnees = list(c("decimal",
                                            "decimal",
                                            "decimal",
                                            "decimal",
                                            "texte"),
                                          c("numerique",
                                            "numerique",
                                            "decimal",
                                            "numerique",
                                            "numerique",
                                            "numerique")),
                largeurs = list(c("auto"), c("auto")),
                format = "primeur",
                col_debut = 2)


## Sauver
saveWorkbook(wb = mon_classeur, file = "tableau.xlsx", overwrite = TRUE)

## Ouvrir
browseURL("tableau.xlsx")
}
