% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_excel_depuis_plan.R
\name{creer_excel_depuis_plan}
\alias{creer_excel_depuis_plan}
\title{Crée un classeur et le formate}
\usage{
creer_excel_depuis_plan(
  plan,
  format,
  sep = ",",
  type_virgule = ",",
  col_debut,
  save = TRUE,
  path = "tableau1.xlsx"
)
}
\arguments{
\item{plan}{fichier .xlsx ou data frame contenant les détails pour chaque tableau}

\item{format}{CHAR le type de publication : "chiffres_et_donnees" ou "primeur"}

\item{sep}{CHAR le caractère utilisé pour séparer les colonnes des fichiers}

\item{type_virgule}{CHAR le type de séparation décimale utilisée : "." ou ","}

\item{col_debut}{NUM la colonne à laquelle les tableaux démarrent}

\item{save}{LGL TRUE si export du classeur au format xlsx}

\item{path}{CHAR chemin vers le tableau à sauvegarder}
}
\value{
un objet de type workbook formaté
}
\description{
Cette fonction permet de formater tous les tableaux d'un classeur selon le type de publication choisi, simplement à partir d'un plan. Les fichiers indiqués dans le plan doivent être dans le répertoire de travail.
}
\examples{
library(agreste)
library(dplyr)
library(openxlsx)
### Import du fichier de plan
plan = agreste::modele_plan

### Création des csv 
agreste::resultat_1 \%>\%
  write.csv(file = "resultat_1.csv", row.names = FALSE)
agreste::resultat_2 \%>\%
  write.csv(file = "resultat_2.csv", row.names = FALSE)
agreste::resultat_3 \%>\%
  write.csv(file = "resultat_3.csv", row.names = FALSE)
agreste::lait_vache \%>\%
  write.xlsx(file = "lait_vache.xlsx", row.names = FALSE)
write.xlsx(plan, "plan.xlsx", row.names = FALSE)
agreste::scieries2020 \%>\%
  write.xlsx(file = "scieries2020.xlsx", row.names = FALSE)
write.xlsx(plan, "plan.xlsx", row.names = FALSE)
### Formatage
workbook <- creer_excel_depuis_plan(plan = "plan.xlsx",
                                    format = "chiffres_et_donnees",
                                    col_debut = 1,
                                    save = TRUE,
                                    path = "document_a_envoyer.xlsx",
                                    type_virgule = ".")

}
