% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formater_auto.R
\name{formater_auto}
\alias{formater_auto}
\title{Formater un classeur entier}
\usage{
formater_auto(
  classeur,
  format,
  liste_feuilles_avec_note,
  liste_type_donnees,
  col_debut = 2
)
}
\arguments{
\item{classeur}{workbook créé avec openXLSX}

\item{format}{CHAR le type de publication : "chiffres_et_donnees" ou "primeur"}

\item{liste_feuilles_avec_note}{vecteur contenant le nom de chaque feuille avec une note de lecture}

\item{liste_type_donnees}{LIST liste de vecteurs contenant les types de données de chaque colonne}

\item{col_debut}{NUM la colonne à laquelle les tableaux démarrent}
}
\value{
Rien n'est renvoyé mais le classeur est modifié
}
\description{
Cette fonction permet de formater tous les tableaux d'un classeur selon le type de publication choisi
}
\examples{
library(openxlsx)
library(agreste)

## Création d'un classeur avec deux feuilles et deux tableaux
mon_classeur <- createWorkbook()

ajouter_tableau_excel(classeur = mon_classeur,
                      tableau = iris,
                      nom_feuille = "iris",
                      ligne_debut = 5,
                      col_debut = 2)

ajouter_titre_tableau(classeur = mon_classeur,
                      nom_feuille = "iris",
                      titre = "Données fleurs iris",
                      col_debut = 2,
                      format = "primeur")
ajouter_source(classeur = mon_classeur,
               nom_feuille = "iris",
               source = "Ceci est une source quelconque",
               format = "primeur",
               col_debut = 2,
               avec_note = FALSE)
ajouter_champ(classeur = mon_classeur,
              nom_feuille = "iris",
              champ = "France métropolitaine",
              format = "primeur",
              col_debut = 2)

ajouter_tableau_excel(classeur = mon_classeur,
                      tableau = airquality,
                      nom_feuille = "airquality",
                      ligne_debut = 5,
                      col_debut = 2)

ajouter_titre_tableau(classeur = mon_classeur,
                      nom_feuille = "airquality",
                      titre = "Données qualité de l'air",
                      col_debut = 2,
                      format = "primeur")
ajouter_note_lecture(classeur = mon_classeur,
                     nom_feuille = "airquality",
                     note = "blablabla",
                     format = "primeur",
                     col_debut = 2)
ajouter_source(classeur = mon_classeur,
               nom_feuille = "airquality",
               source = "Eurostat",
               format = "primeur",
               col_debut = 2,
               avec_note = TRUE)
ajouter_champ(classeur = mon_classeur,
              nom_feuille = "airquality",
              champ = "Daily air quality measurements in New York, May to September 1973.",
              format = "primeur",
              col_debut = 2)

## Formater comme un Chiffres et Données
formater_auto(mon_classeur,
              format = "primeur",
              liste_feuilles_avec_note = c("airquality"),
              liste_type_donnees = list(c("decimal",
                                          "decimal",
                                          "decimal",
                                          "decimal",
                                          "texte"),
                                        c("numerique",
                                          "numerique",
                                          "decimal",
                                          "numerique",
                                          "numerique",
                                          "numerique")),
              col_debut = 2)

## Sauver
saveWorkbook(wb = mon_classeur, file = "tableau.xlsx", overwrite = TRUE)

## Ouvrir
browseURL("tableau.xlsx")

}
