% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formater_auto.R
\name{formater_auto}
\alias{formater_auto}
\title{Formater un classeur entier}
\usage{
formater_auto(
  classeur,
  format,
  liste_feuilles_avec_note,
  liste_type_donnees,
  liste_lignes_titre,
  liste_lignes_section,
  liste_lignes_sous_total,
  liste_lignes_precision_1,
  liste_lignes_precision_2,
  liste_lignes_precision_3,
  liste_lignes_precision_4,
  liste_lignes_total,
  liste_unif_unites,
  liste_cellules_fusion,
  type_virgule = ",",
  col_debut = 2,
  avec_titre = FALSE
)
}
\arguments{
\item{classeur}{workbook créé avec openXLSX}

\item{format}{CHAR le type de publication :
"chiffres_et_donnees" ou "primeur"}

\item{liste_feuilles_avec_note}{vecteur contenant le nom de chaque
feuille avec une note de lecture}

\item{liste_type_donnees}{LIST liste de vecteurs contenant les types de
données de chaque colonne}

\item{liste_lignes_titre}{LIST liste des lignes de titres de colonnes}

\item{liste_lignes_section}{LIST liste de vecteurs contenant les numéros des
lignes du tableau qui sont des sections}

\item{liste_lignes_sous_total}{LIST liste des vecteurs contenant les numéros
des lignes du tableau qui sont des sous-totaux}

\item{liste_lignes_precision_1}{LIST liste des vecteurs contenant les numéros
des lignes du tableau qui sont des lignes de précision de niveau 1}

\item{liste_lignes_precision_2}{LIST liste des vecteurs contenant les numéros
des lignes du tableau qui sont des lignes de précision de niveau 2}

\item{liste_lignes_precision_3}{LIST liste des vecteurs contenant les numéros
des lignes du tableau qui sont des lignes de précision de niveau 3}

\item{liste_lignes_precision_4}{LIST liste des vecteurs contenant les numéros
des lignes du tableau qui sont des lignes de précision de niveau 4}

\item{liste_lignes_total}{LIST liste des vecteurs contenant les numéros
des lignes du tableau qui sont des lignes de totaux}

\item{liste_unif_unites}{VEC un vecteur de valeurs logiques pour indiquer
les feuilles ayant des unités de colonnes différentes}

\item{liste_cellules_fusion}{LIST la liste des cellules à fusionner}

\item{col_debut}{NUM la colonne à laquelle les tableaux démarrent}

\item{avec_titre}{LGL si le classeur est un primeur avec un titre}
}
\value{
Rien n'est renvoyé mais le classeur est modifié
}
\description{
Cette fonction permet de formater tous les tableaux d'un classeur selon le type de publication choisi
}
\examples{
library(openxlsx)
library(agreste)

## Création d'un classeur avec deux feuilles et deux tableaux
mon_classeur <- createWorkbook()

ajouter_tableau_excel(classeur = mon_classeur,
                      tableau = iris,
                      nom_feuille = "iris",
                      ligne_debut = 5,
                      col_debut = 2)

ajouter_titre_tableau(classeur = mon_classeur,
                      nom_feuille = "iris",
                      titre = "Données fleurs iris",
                      col_debut = 2,
                      format = "primeur")
ajouter_source(classeur = mon_classeur,
               nom_feuille = "iris",
               source = "Ceci est une source quelconque",
               format = "primeur",
               col_debut = 2,
               avec_note = FALSE,
               avec_titre = FALSE)
ajouter_champ(classeur = mon_classeur,
              nom_feuille = "iris",
              champ = "France métropolitaine",
              format = "primeur",
              col_debut = 2,
              avec_titre = FALSE)

ajouter_tableau_excel(classeur = mon_classeur,
                      tableau = airquality,
                      nom_feuille = "airquality",
                      ligne_debut = 5,
                      col_debut = 2)

ajouter_titre_tableau(classeur = mon_classeur,
                      nom_feuille = "airquality",
                      titre = "Données qualité de l'air",
                      col_debut = 2,
                      format = "primeur")
ajouter_note_lecture(classeur = mon_classeur,
                     nom_feuille = "airquality",
                     note = "blablabla",
                     format = "primeur",
                     col_debut = 2,
                     avec_titre = FALSE)
ajouter_source(classeur = mon_classeur,
               nom_feuille = "airquality",
               source = "Eurostat",
               format = "primeur",
               col_debut = 2,
               avec_note = TRUE,
               avec_titre = FALSE)
ajouter_champ(classeur = mon_classeur,
              nom_feuille = "airquality",
              champ = "Daily air quality measurements in New York, May to September 1973.",
              format = "primeur",
              col_debut = 2,
              avec_titre = FALSE)

ajouter_titre_primeur(classeur = mon_classeur,
                      titre = "Exemple 2022")

taille_colonnes(classeur = mon_classeur,
                liste_type_donnees = list(c("decimal",
                                            "decimal",
                                            "decimal",
                                            "decimal",
                                            "texte"),
                                          c("numerique",
                                            "numerique",
                                            "decimal",
                                            "numerique",
                                            "numerique",
                                            "numerique")),
                largeurs = list(c("auto"), c("auto")),
                col_debut = 2)

## Formater comme un Primeur
formater_auto(classeur = mon_classeur,
              format = "primeur",
              liste_feuilles_avec_note = c("airquality"),
              liste_type_donnees = list(c("decimal",
                                          "decimal",
                                          "decimal",
                                          "decimal",
                                          "texte"),
                                        c("numerique",
                                          "numerique",
                                          "decimal",
                                          "numerique",
                                          "numerique",
                                          "numerique")), 
              liste_lignes_titre = list(c(1), c(1)),
              liste_lignes_section = list(c(5, 9),
                                          c(3, 8)),
              liste_lignes_sous_total = list(c(4, 8),
                                             c(2, 7)),
              liste_lignes_precision_1 = list(c(2,3),
                                            c(0)),
              liste_lignes_precision_2 = list(c(0), c(0)),
              liste_lignes_precision_3 = list(c(0), c(0)),
              liste_lignes_precision_4 = list(c(0), c(0)),
              liste_lignes_total = list(c(150),
                                        c(100)),
              liste_unif_unites = c(TRUE, TRUE),
              liste_cellules_fusion = list(c(0), c(0)),
              col_debut = 2,
              avec_titre = TRUE)

## Sauver
saveWorkbook(wb = mon_classeur, file = "tableau.xlsx", overwrite = TRUE)

## Ouvrir
browseURL("tableau.xlsx")

}
