% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geler.R
\name{geler}
\alias{geler}
\title{Geler une(des) ligne(s) et/ou une(des) colonne(s)}
\usage{
geler(classeur, nom_feuille, ligne = 0, col = 0)
}
\arguments{
\item{classeur}{workbook créé avec openXLSX}

\item{nom_feuille}{nom de la feuille}

\item{ligne}{la ligne à geler (0 si pas de ligne à geler)}

\item{col}{la colonne à geler (0 si pas de colonne à geler)}
}
\value{
Rien n'est renvoyé mais la feuille du classeur est modifiée
}
\description{
Cette fonction permet de geler une ou plusieurs lignes ainsi qu'une ou plusieurs colonnes d'une feuille de classeur
}
\examples{

library(openxlsx)
library(agreste)

# Création d'un classeur avec une feuille contenant iris 
mon_classeur <- createWorkbook()
ajouter_tableau_excel(classeur = mon_classeur, tableau = iris, nom_feuille = "iris", ligne_debut = 3, col_debut = 2)

# gel 
geler(mon_classeur, "iris", ligne = 3, col = 1)

# Sauver le classeur
saveWorkbook(wb = mon_classeur, file = "tableau.xlsx", overwrite = TRUE)
# ouverture
browseURL("tableau.xlsx")
}
