% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_excel_depuis_plan.R
\name{read_any_file}
\alias{read_any_file}
\title{Lit un fichier}
\usage{
read_any_file(
  filename,
  header = TRUE,
  sep = ",",
  decimal_mark = ".",
  sheet = 1
)
}
\arguments{
\item{filename}{CHAR le nom du fichier, extension incluse}
}
\value{
A dataframe
}
\description{
Prend en charge les fichiers csv, xlsx, rds
}
